Move custom logging module
