+++ pdfmerger/custom/loggingC.py
